testy_v2 przeniesienie testu_v2 do nowej klasy
